#620: Move cli init so it happens directly in the cli entrypoint
