+++ lib/bootstrap.js
@@ -11,9 +11,9 @@
  * // Initialize Lando in CLI mode
  * bootstrap({mode: 'cli'})
  *
- * // Initialize CLI
+ * // Initialize some other thing
  * .tap(function(lando) {
- *   return lando.cli.init(lando);
+ *   return cli.init(lando);
  * })
  */
 
@@ -96,7 +96,7 @@ module.exports = _.once(function(opts) {
   // Return our plugins so we can init them
   .then(function() {
     lando.log.verbose('Trying to load plugins', lando.config.plugins);
-    return lando.config.plugins;
+    return lando.config.plugins || [];
   })
 
   // Init the plugins
