+++ lib/cli.js
@@ -1,12 +1,12 @@
 /**
- * Contains methods to help initialize and display the CLI
+ * Contains some CLI artwork and formatting helpers
  *
  * @since 3.0.0
  * @module cli
  * @example
  *
- * // Initialize CLI
- * return lando.cli.init(lando);
+ * // Get the start header
+ * return lando.cli.startHeader();
  */
 
 'use strict';
@@ -15,16 +15,7 @@
 var _ = require('./node')._;
 var chalk = require('./node').chalk;
 var os = require('os');
-var path = require('path');
 var Table = require('cli-table');
-var sudoBlock = require('sudo-block');
-
-// Yargonaut must come before yargs
-var yargonaut = require('yargonaut');
-yargonaut.style('green').errorsStyle('red');
-
-// Get yargs
-var yargs = require('yargs');
 
 /**
  * Returns a cheeky header that can be used after an app is started.
@@ -185,99 +176,5 @@ exports.Table = function(opts) {
 
   // Return our default table
   return new Table(_.merge(tableDefaults, opts));
-
-};
-
-/**
- * Initializes the CLI.
- *
- * This will either print the CLI usage to the console or route the command and
- * options given by the user to the correct place.
- *
- * @since 3.0.0
- * @fires pre-cli-load
- * @param {Object} lando - An initialized lando object.
- * @example
- *
- * // Initialize the CLI
- * return lando.cli.init(lando);
- */
-exports.init = function(lando) {
-
-  // Log
-  lando.log.info('Initializing cli');
-
-  // Get global tasks
-  var tasks = _.sortBy(lando.tasks.tasks, 'name');
-
-  // Get cmd
-  var cmd = '$0';
-
-  // If we are packaged lets get something else
-  if (_.has(process, 'pkg')) {
-    cmd = path.basename(_.get(process, 'execPath', 'lando'));
-  }
-
-  // Define our default CLI
-  yargs
-    .usage('Usage: ' + cmd + ' <command> [args] [options] [-- global options]');
-
-  /**
-   * Event that allows other things to alter the tasks being loaded to the CLI.
-   *
-   * @since 3.0.0
-   * @event module:cli.event:pre-cli-load
-   * @property {Object} tasks An object of Lando tasks
-   * @example
-   *
-   * // As a joke remove all tasks and give us a blank CLI
-   * lando.events.on('pre-cli-load', function(tasks) {
-   *   tasks = {};
-   * });
-   */
-  return lando.events.emit('pre-cli-load', tasks)
-
-  // No sudo!
-  .then(function() {
-    sudoBlock(lando.node.chalk.red('Lando should never be run as root!'));
-  })
-
-  // Print our result
-  .then(function() {
-
-    // Parse any global opts for usage later
-    tasks.largv = lando.tasks.parseGlobals();
-
-    // Create epilogue for our global options
-    var epilogue = [
-      chalk.green('Global Options:\n'),
-      '  --help, -h  Show help\n',
-      '  --verbose, -v, -vv, -vvv, -vvvv  Change verbosity of output'
-    ];
-
-    // Loop through the tasks and add them to the CLI
-    _.forEach(tasks, function(task) {
-      lando.log.verbose('Loading cli task %s', task.name);
-      yargs.command(lando.tasks.parseToYargs(task));
-    });
-
-    // Invoke help if global option is specified
-    if (tasks.largv.help) {
-      yargs.showHelp();
-      process.exit(0);
-    }
-
-    // Finish up the yargs
-    var argv = yargs
-      .strict(true)
-      .demandCommand(1, 'You need at least one command before moving on')
-      .epilog(epilogue.join(''))
-      .wrap(yargs.terminalWidth() * 0.75)
-      .argv;
-
-    // Log the CLI
-    lando.log.debug('CLI args', argv);
-
-  });
 
 };
